initial split of common utilities into separate files
