+++ src/SessionManager/SessionManager.h
@@ -46,6 +46,8 @@ private:
     std::string _auth_token;
     std::shared_ptr<FileListHandler> _file_list_handler;
     std::shared_ptr<CartaGrpcService> _grpc_service;
+    
+    std::string IPAsText(std::string_view binary);
 };
 } // namespace carta
 #endif // CARTA_BACKEND_SRC_SESSIONMANAGER_SESSIONMANAGER_H_
