+++ src/Util/String.h
@@ -1,0 +1,7 @@
+#ifndef CARTA_BACKEND__UTIL_STRING_H_
+#define CARTA_BACKEND__UTIL_STRING_H_
+
+// split input string into a vector of strings by delimiter
+void SplitString(std::string& input, char delim, std::vector<std::string>& parts);
+
+#endif // CARTA_BACKEND__UTIL_STRING_H_
